It works, it works! (no library loaded, but everything else but marshalling
